move assessment/api/views to assessment/api
This brings assessment api views inline with other app api views.
