+++ project/epimeta/api.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from assessment.api.views import AssessmentViewset
+from assessment.api import AssessmentViewset
 
 from . import models, serializers
