+++ project/study/api.py
@@ -6,7 +6,7 @@ from rest_framework import viewsets
 from rest_framework.decorators import list_route
 from rest_framework.response import Response
 
-from assessment.api.views import (
+from assessment.api import (
     AssessmentLevelPermissions, InAssessmentFilter, DisabledPagination)
 
 from . import models, serializers
