+++ project/summary/api.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 
-from assessment.api.views import AssessmentViewset, DisabledPagination
+from assessment.api import AssessmentViewset, DisabledPagination
 
 from . import models, serializers
